% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aupc_pscore.R
\name{plot_aupc_pscore}
\alias{plot_aupc_pscore}
\title{Compare AUPC and P-score treatment rankings}
\usage{
plot_aupc_pscore(aupc, pscore, outcome_name = "")
}
\arguments{
\item{aupc}{A data frame returned by \code{aupc()}.}

\item{pscore}{A \code{pscore_civs} object.}

\item{outcome_name}{Optional title for the plot depicting the outcome name.}
}
\value{
A \code{ggplot2} object.
}
\description{
Compare treatment rankings derived from AUPC and extended P-scores.
}
\details{
Compare AUPC and P-score treatment rankings
}
