% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscore_civs.R
\name{print.pscore_civs}
\alias{print.pscore_civs}
\title{Print a pscore_civs object}
\usage{
\method{print}{pscore_civs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pscore_civs}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Prints the \code{all} component of a \code{pscore_civs} object.
}
