% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{PscoresPlus-package}
\alias{PscoresPlus-package}
\alias{PscoresPlus}
\title{PscoresPlus: Extended P-scores for Network Meta-Analysis}
\description{
PscoresPlus extends the P-score framework for ranking treatments in Network
Meta-Analysis (NMA) by incorporating Clinically Important Values (CIVs).
}
\keyword{internal}
