\name{aupc}
\alias{aupc}
\title{Compute Area Under the P-score Curve (AUPC) from \code{pscore_civs} output}
\description{
Computes the Area Under the P-score Curve (AUPC) for each treatment, using the
P-scores and CIV combinations generated by \code{pscore_civs}. AUPC is
interpreted as the average probability that a treatment beats all others,
averaged over the specified grid of CIVs.
}
\usage{
aupc(x, excel = FALSE)
}
\arguments{
  \item{x}{A \code{pscore_civs} object (output of \code{pscore_civs} function).}
  \item{excel}{Logical. If \code{TRUE}, the results are exported to an Excel file.
  Default is \code{FALSE}.}
}
\value{
A data frame with the following columns:
  \item{AUPC}{Numeric vector of Area Under the P-score Curve (AUPC) values for each
    treatment, interpreted as the average probability that the treatment beats
    all others over the considered CIV combinations.}
  \item{Treatment}{Treatment labels}
  \item{ranking}{Ranking of treatments based on AUPC, with highest AUPC ranked 1
    (top-ranked treatment).}
}
\details{
AUPC is computed from the P-scores in a \code{pscore_civs} object and used to rank
treatments over the specified CIV grid (i.e., all combinations of the specified CIV values).

If \code{excel = TRUE}, the results are also exported to an Excel file in the working directory. Default is \code{FALSE}.
}
\examples{
\dontrun{

# Import datasets (netmeta objects)

data("efficacy")
data("weight_gain")

#Efficacy
#Calculating AUPC for outcome efficacy with a grid of 50 values between 0 and 0.5.

out_eff <- pscore_civs(
  list(efficacy),
  CIVs = list(seq(0, 0.5, length.out = 50)),
  correlation = NULL,
  type = c("H")
)

aupc(out_eff)
# Export results to Excel
aupc(out_eff, excel = TRUE)

#Weight Gain
#Calculating AUPC for outcome weight gain with a grid of 50 values between -0.5 and 0.

out_wg<- pscore_civs(
  list(weight_gain),
  CIVs = list(seq(-0.5, 0, length.out = 50)),
  correlation = NULL,
  type = c("H")
)

aupc(out_wg)
# Export results to Excel
aupc(out_wg, excel = TRUE)

#Efficacy and Weight Gain Outcomes
# Run pscores_civs with a grid of 50 values between 0 and 0.5 for efficacy and -0.5 and 0 for weight gain.

#correlation matrix between the two outcomes (the assumed correlation between efficacy and weight gain is −0.5)
cor_mat <- matrix(c(1, -0.5,
                   -0.5, 1),
                    nrow = 2,
                    byrow = TRUE)

out <- pscore_civs(
  list(efficacy, weight_gain),
  CIVs = list(seq(0, 0.5, length.out = 50),
              seq(-0.5, 0, length.out = 50)),
  correlation = cor_mat,
  type = c("H", "H")
)

aupc(out)

# Export results to Excel
aupc(out, excel = TRUE)

}
}
